Reverting changes to DTOs to re-enable foreign key constraints. Adding temporary condition to executing database schema creation code to skip foreign key constraints for sql ce.

+++ src/Umbraco.Core/Models/Rdbms/AppTreeDto.cs
@@ -17,7 +17,7 @@ namespace Umbraco.Core.Models.Rdbms
         public bool Initialize { get; set; }
 
         [Column("treeSortOrder")]
-        public int SortOrder { get; set; }
+        public byte SortOrder { get; set; }
 
         [Column("appAlias")]
         [PrimaryKeyColumn(AutoIncrement = false, Clustered = true, Name = "PK_umbracoAppTree", OnColumns = "[appAlias], [treeAlias]")]
